Move the signal code out of the models and into model/signals
