+++ src/sentry/workflow_engine/apps.py
@@ -5,6 +5,7 @@ class Config(AppConfig):
     name = "sentry.workflow_engine"
 
     def ready(self) -> None:
-        # Import our base DataConditionHandlers for the workflow engine platform
+        # Import items that use registries or respond to events
         import sentry.workflow_engine.handlers  # NOQA
+        import sentry.workflow_engine.models.signals  # NOQA
         from sentry.workflow_engine.endpoints import serializers  # NOQA
