+++ src/sentry/workflow_engine/models/data_source.py
@@ -4,14 +4,11 @@ import logging
 from typing import Generic, TypeVar
 
 from django.db import models
-from django.db.models.signals import pre_save
-from django.dispatch import receiver
 
 from sentry.backup.dependencies import NormalizedModelName, PrimaryKeyMap
 from sentry.backup.helpers import ImportFlags
 from sentry.backup.scopes import ImportScope, RelocationScope
 from sentry.db.models import DefaultFieldsModel, FlexibleForeignKey, region_silo_model
-from sentry.utils.registry import NoRegistrationExistsError
 from sentry.workflow_engine.models.data_source_detector import DataSourceDetector
 from sentry.workflow_engine.registry import data_source_type_registry
 from sentry.workflow_engine.types import DataSourceTypeHandler
@@ -90,19 +87,3 @@ class DataSource(DefaultFieldsModel):
             return None
 
         return old_pk
-
-
-@receiver(pre_save, sender=DataSource)
-def ensure_type_handler_registered(sender, instance: DataSource, **kwargs):
-    """
-    Ensure that the type of the data source is valid and registered in the data_source_type_registry
-    """
-    data_source_type = instance.type
-
-    if not data_source_type:
-        raise ValueError(f"No group type found with type {instance.type}")
-
-    try:
-        data_source_type_registry.get(data_source_type)
-    except NoRegistrationExistsError:
-        raise ValueError(f"No data source type found with type {data_source_type}")
