+++ src/sentry/workflow_engine/models/detector.py
@@ -7,8 +7,6 @@ from typing import TYPE_CHECKING, Any, ClassVar, TypedDict
 
 from django.conf import settings
 from django.db import models
-from django.db.models.signals import pre_save
-from django.dispatch import receiver
 from jsonschema import ValidationError
 
 from sentry.backup.scopes import RelocationScope
@@ -231,11 +229,3 @@ def enforce_config_schema(instance: Detector) -> None:
         raise ValidationError("Detector config must be a dictionary")
 
     instance.validate_config(group_type.detector_settings.config_schema)
-
-
-@receiver(pre_save, sender=Detector)
-def enforce_config_schema_signal(sender, instance: Detector, **kwargs):
-    """
-    This needs to be a signal because the grouptype registry's entries are not available at import time.
-    """
-    enforce_config_schema(instance)
