+++ src/sentry/workflow_engine/models/signals/__init__.py
@@ -1,0 +1,5 @@
+from .action import *  # NOQA
+from .data_condition import *  # NOQA
+from .data_source import *  # NOQA
+from .detector import *  # NOQA
+from .workflow import *  # NOQA
